wasmserve: graceful shutdown on sigint (#33)
The wasmserver program creates a temporary directory to store the built main.wasm binary but does not clean up the
temporary directory on shutdown. This change catches ctrl-c and performs a clean shutdown. The changes are roughly

1. call `os.RemoveAll(tmpdir)` in `main()`
2. use a non-default http.Server so that it can be shutdown via `server.Shutdown()`
3. use `os.Signal` to catch sigint

Also I added some extra minor logging so the behavior of the server was more apparent.